# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger Test Skill Laravel Pleno",
        "description": "",
        "version": "1.0.0"
    },
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "Register"
                ],
                "summary": "Register a new user",
                "operationId": "register",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "display_name": {
                                        "type": "string"
                                    },
                                    "full_name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "document": {
                                        "type": "string"
                                    },
                                    "document_type": {
                                        "type": "string"
                                    },
                                    "phone": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "Jessica",
                                    "full_name": "Smith",
                                    "email": "jessica@email.com",
                                    "password": "123",
                                    "document": "864.997.710-31",
                                    "document_type": "cpf"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Client"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Login"
                ],
                "summary": "create a token to have access to the resources",
                "operationId": "login",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "email": "jessica@email.com",
                                    "password": "123"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Client"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/client": {
            "get": {
                "tags": [
                    "client"
                ],
                "summary": "Fetch a listing of clients registed",
                "operationId": "index",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Client"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/client/{id}": {
            "get": {
                "tags": [
                    "client"
                ],
                "summary": "Fetch a client by its id",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "client id",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Client not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "client"
                ],
                "summary": "Update a client by its id",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "client id",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "display_name": {
                                        "type": "string"
                                    },
                                    "full_name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "document": {
                                        "type": "string"
                                    },
                                    "document_type": {
                                        "type": "string"
                                    },
                                    "phone": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "Jessica",
                                    "full_name": "Smith",
                                    "email": "jessica@email.com",
                                    "password": "123",
                                    "document": "864.997.710-31",
                                    "document_type": "cpf"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "successful operation"
                    },
                    "404": {
                        "description": "Client not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "client"
                ],
                "summary": "Delete a client by its id",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "client id",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "successful operation"
                    },
                    "404": {
                        "description": "Client not found"
                    }
                }
            }
        },
        "/api/product": {
            "get": {
                "tags": [
                    "product"
                ],
                "summary": "Fetch a listing of products registed",
                "operationId": "index",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Product"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/product/{id}": {
            "get": {
                "tags": [
                    "product"
                ],
                "summary": "Fetch a product by its id",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "product id",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "product"
                ],
                "summary": "Update a product by its id",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "product id",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "price": {
                                        "type": "float"
                                    },
                                    "sku": {
                                        "type": "string"
                                    },
                                    "stock_amount": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "sapato",
                                    "description": "Sapato exclusivo da Linha Reserva pra Calçar!",
                                    "price": 299.5,
                                    "sku": "TSH-MED-WHI-COT",
                                    "stock_amount": 10
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "404": {
                        "description": "Product not found"
                    }
                }
            },
            "post": {
                "tags": [
                    "product"
                ],
                "summary": "Register a new product",
                "operationId": "store",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "price": {
                                        "type": "float"
                                    },
                                    "sku": {
                                        "type": "string"
                                    },
                                    "stock_amount": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "sapato",
                                    "description": "Sapato exclusivo da Linha Reserva pra Calçar!",
                                    "price": 299.5,
                                    "sku": "TSH-MED-WHI-COT",
                                    "stock_amount": 10
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "successful operation"
                    }
                }
            },
            "delete": {
                "tags": [
                    "product"
                ],
                "summary": "Delete a product by its id",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "product id",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "successful operation"
                    },
                    "404": {
                        "description": "Product not found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Client": {
                "title": "Client model",
                "description": "Client model",
                "required": [
                    "display_name",
                    "full_name",
                    "email",
                    "document",
                    "document_type"
                ],
                "xml": {
                    "name": "Client"
                }
            },
            "Product": {
                "title": "Product model",
                "description": "Product model",
                "required": [
                    "name",
                    "description",
                    "price",
                    "sku",
                    "client_id"
                ],
                "xml": {
                    "name": "Product"
                }
            }
        }
    },
    "security": [
        []
    ]
}